Move CompletableFuture tests to scala_2.12 suites
